requirements.txt
# Libreria per la gestione dei file di configurazione YAML
pyyaml==6.0.1

# Opzionale: per compilare il client in un file .exe (Windows)
# pyinstaller==6.4.0